(hel-401) Export compte de résultat view model
